Add basic reporting in the plugin settings screen.
- Date picker on the settings page.
- Depending on the dates, we calculate the number of commits for that period, as well as a number of actions.

Actions are all hardcoded right now, will need to improve that, and add more.
I will also need to take plural / singular forms into account in the display.

It might also be good to include a save button to save a snapshot for a specific date range, and have the saved snapshot added to a graph.