Move chart building functions.
Now that we don't use the charts in the admin, we can consolidate where all our functions live.
Shortcodes will only live in the shortcodes directory.
We also do not need the date picking feature anymore since we do not have an interface to pick a report's date anymore. It all happens via shortcode paramters